removed dependency on the graphql from framework
- this commit changes the PR #2516
