+++ packages/framework/src/Component/HttpFoundation/TransactionalMasterRequestListener.php
@@ -5,9 +5,6 @@ declare(strict_types=1);
 namespace Shopsys\FrameworkBundle\Component\HttpFoundation;
 
 use Doctrine\ORM\EntityManagerInterface;
-use GraphQL\Error\SyntaxError;
-use GraphQL\Language\AST\OperationDefinitionNode;
-use GraphQL\Language\Parser;
 use Symfony\Component\HttpKernel\Event\ExceptionEvent;
 use Symfony\Component\HttpKernel\Event\RequestEvent;
 use Symfony\Component\HttpKernel\Event\ResponseEvent;
@@ -17,9 +14,6 @@ class TransactionalMasterRequestListener
     protected bool $inTransaction = false;
 
     protected static bool $isManuallyRollbacked = false;
-
-    protected const GRAPHQL_ENDPOINT_ROUTE = 'overblog_graphql_endpoint';
-    protected const QUERY_TYPE = 'query';
 
     /**
      * @param \Doctrine\ORM\EntityManagerInterface $em
@@ -39,7 +33,7 @@ class TransactionalMasterRequestListener
      */
     public function onKernelRequest(RequestEvent $event): void
     {
-        if ($event->isMainRequest() && !$this->inTransaction && !$this->isRequestGraphQlQuery($event)) {
+        if ($event->isMainRequest() && !$this->inTransaction) {
             $this->em->beginTransaction();
             $this->inTransaction = true;
         }
@@ -65,44 +59,5 @@ class TransactionalMasterRequestListener
             $this->em->rollback();
             $this->inTransaction = false;
         }
-    }
-
-    /**
-     * @param \Symfony\Component\HttpKernel\Event\RequestEvent $requestEvent
-     * @return bool
-     */
-    protected function isRequestGraphQlQuery(RequestEvent $requestEvent): bool
-    {
-        if ($requestEvent->getRequest()->attributes->get('_route') !== static::GRAPHQL_ENDPOINT_ROUTE) {
-            return false;
-        }
-
-        $requestContent = $requestEvent->getRequest()->getContent();
-
-        if ($requestContent === null || $requestContent === '') {
-            return false;
-        }
-
-        $source = json_decode($requestContent, true);
-
-        if (!array_key_exists(static::QUERY_TYPE, $source)) {
-            return false;
-        }
-
-        $queryString = $source[static::QUERY_TYPE];
-
-        try {
-            $parsed = Parser::parse($queryString);
-
-            foreach ($parsed->definitions as $definition) {
-                if ($definition instanceof OperationDefinitionNode) {
-                    return $definition->operation === static::QUERY_TYPE;
-                }
-            }
-        } catch (SyntaxError) {
-            return false;
-        }
-
-        return false;
     }
 }
